Refactored adminPage
